make runIncrementalCompile.sh take cabal-dev sandbox as a parameter

+++ build-scripts/runIncrementalCompile.sh
@@ -1,13 +1,18 @@
 #!/bin/bash
 
+if [ -z "$1" ]; then
+  echo "Usage: $0 <cabal-dev sandbox directory>"
+  exit
+fi
+
 rm -f kontrakcja-test.tix
-cabal install --only-dependencies
-cabal configure -ftest-coverage
-cabal build --ghc-options="-O0 -c"
+cabal-dev --sandbox="$1" install --only-dependencies
+cabal-dev --sandbox="$1" configure -ftest-coverage
+cabal-dev --sandbox="$1" build --ghc-options="-O0 -c"
 
 if [ "$TEAMCITY_VERSION" = "" ]; then
-  cabal build --ghc-options="-O0 -optl -O0"
+  cabal-dev --sandbox="$1" build --ghc-options="-O0 -optl -O0"
 else
-  cabal build --ghc-options="-O0 -optl -O0" 2>&1 | runghc build-scripts/Teamcity.hs ghc
+  cabal-dev --sandbox="$1" build --ghc-options="-O0 -optl -O0" 2>&1 | runghc build-scripts/Teamcity.hs ghc
   exit "${PIPESTATUS[0]}"
 fi
